updating color descrpiption fix

+++ src/main/java/com/source/loader/model3d/camera/point/dto/CameraPointDTO.java
@@ -17,17 +17,17 @@ import java.util.UUID;
 @NoArgsConstructor
 public class CameraPointDTO implements Mapper<CameraPointDTO, CameraPoint> {
 
-    private Float camera_z_position;
+    private Double camera_z_position;
 
-    private Float camera_y_position;
+    private Double camera_y_position;
 
-    private Float camera_x_position;
+    private Double camera_x_position;
 
-    private Float point_z_position;
+    private Double point_z_position;
 
-    private Float point_y_position;
+    private Double point_y_position;
 
-    private Float point_x_position;
+    private Double point_x_position;
 
     private String description;
